ref. [issuer] 디렉터리 구조 재설정
