+++ license-issuer/src/main/java/com/licenseissuer/model/dto/TempLicenseDto.java
@@ -1,6 +1,6 @@
 package com.licenseissuer.model.dto;
 
-import com.licenseissuer.model.LicenseType;
+import com.licenseissuer.model.enums.LicenseType;
 import lombok.Builder;
 import lombok.Data;
