+++ license-issuer/src/main/java/com/licenseissuer/model/dto/request/LicenseIssueRequest.java
@@ -1,25 +1,23 @@
-package com.licenseissuer.model.dto;
+package com.licenseissuer.model.dto.request;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * 라이센스 업데이트 요청 정보
+ * 라이센스 발급 요청 정보
  */
 @Slf4j
 @Getter
-public class LicenseUpdateRequest {
+public class LicenseIssueRequest {
 	protected String operation;
 	protected String projectName;
 	protected String ipAddress;
 	protected String expDate;
 	protected String issuer;
+	protected String issuerIp;
 
-	protected String processor;
-	protected String processorIp;
-
-	public void setProcessorIp(String processorIp) {
-		this.processorIp = processorIp;
+	public void setIssuerIp(String issuerIp) {
+		this.issuerIp = issuerIp;
 	}
