+++ license-issuer/src/main/java/com/licenseissuer/model/enums/LicenseProcessType.java
@@ -1,4 +1,4 @@
-package com.licenseissuer.model;
+package com.licenseissuer.model.enums;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
